Rename and move some classes.
